+++ pywt/__init__.py
@@ -13,6 +13,7 @@ from __future__ import division, print_function, absolute_import
 
 
 from ._extensions._pywt import *
+from ._extensions._dwt import dwt_coeff_len, dwt_max_level
 from ._functions import *
 from ._multilevel import *
 from ._multidim import *
